# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Crowned In His Glory 2026 Event API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Ticket {
  id: number;
  ticketNumber: string;
  firstName: string;
  lastName: string;
  phone: string;
  gender: string;
  /** @nullable */
  churchAssembly?: string | null;
  status: string;
  seatNumber: string;
  qrCode: string;
  checkedIn?: boolean;
  /** @nullable */
  checkedInAt?: string | null;
  createdAt: string;
}

export interface TicketListResponse {
  tickets: Ticket[];
  total: number;
  page: number;
  limit: number;
}

export interface TicketRegistration {
  /** @minLength 1 */
  firstName: string;
  /** @minLength 1 */
  lastName: string;
  /** @minLength 1 */
  phone: string;
  gender: string;
  churchAssembly?: string;
  quantity?: number;
}

export interface TicketUpdate {
  status?: string;
  checkedIn?: boolean;
}

export interface CheckInInput {
  ticketNumber: string;
  volunteerName?: string;
}

export interface CheckInResult {
  status: string;
  message: string;
  ticket?: Ticket;
}

export interface CheckIn {
  id: number;
  ticketId: number;
  ticketNumber: string;
  attendeeName: string;
  checkedInAt: string;
  /** @nullable */
  volunteerName?: string | null;
}

export interface EventStats {
  totalTickets: number;
  totalRevenue: number;
  checkedIn: number;
  remaining: number;
  pendingCount: number;
  maleCount: number;
  femaleCount: number;
}

export interface AdminCredentials {
  username: string;
  password: string;
}

export interface AdminAuthPayload {
  token: string;
  username: string;
}

export interface AdminVerifyPayload {
  valid: boolean;
  username: string;
}

export type ListTicketsParams = {
status?: string;
search?: string;
page?: number;
limit?: number;
};

export type GetRecentRegistrationsParams = {
limit?: number;
};