# lib/api-zod/src/generated/types/checkInInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Crowned In His Glory 2026 Event API
 * OpenAPI spec version: 0.1.0
 */

export interface CheckInInput {
  ticketNumber: string;
  volunteerName?: string;
}