# lib/api-zod/src/generated/types/checkInResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Crowned In His Glory 2026 Event API
 * OpenAPI spec version: 0.1.0
 */
import type { Ticket } from './ticket';

export interface CheckInResult {
  status: string;
  message: string;
  ticket?: Ticket;
}