# lib/api-zod/src/generated/types/ticket.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Crowned In His Glory 2026 Event API
 * OpenAPI spec version: 0.1.0
 */

export interface Ticket {
  id: number;
  ticketNumber: string;
  firstName: string;
  lastName: string;
  phone: string;
  gender: string;
  /** @nullable */
  churchAssembly?: string | null;
  status: string;
  seatNumber: string;
  qrCode: string;
  checkedIn?: boolean;
  /** @nullable */
  checkedInAt?: string | null;
  createdAt: string;
}